fs/nxffs/nxffs_write.c: Fix a syslog format

+++ fs/nxffs/nxffs_write.c
@@ -41,6 +41,7 @@
 
 #include <nuttx/config.h>
 
+#include <stdint.h>
 #include <string.h>
 #include <fcntl.h>
 #include <crc32.h>
@@ -531,7 +532,7 @@ ssize_t nxffs_write(FAR struct file *filep, FAR const char *buffer,
   ssize_t total;
   int ret;
 
-  finfo("Write %d bytes to offset %d\n", buflen, filep->f_pos);
+  finfo("Write %zd bytes to offset %jd\n", buflen, (intmax_t)filep->f_pos);
 
   /* Sanity checks */
